feat(space): asteroid/kuiper belts, moon dimension + hollow-core traversal, timescale, fixes
Belts:
- BeltDef + belt.fsh: ray/ecliptic-plane intersection, two-layer cellular
  scatter (dust + sparse rocks), Kirkwood-gap clumping, slow drift
- Asteroid belt (3200-5400, gray) + Kuiper belt (50000-68000, icy)

Moon dimension (first landable planet):
- MOON dim keys + datagen (biome/type/levelstem); MoonDimensionEffects
- MoonChunkGenerator: FBM hills + craters; HOLLOW structure (crust 70-20,
  hollow interior, glowing core -40 to bottom)
- New blocks: moon_regolith/stone/deepstone/core (Python procedural textures,
  clean low-freq shading to avoid visual noise)
- moon's PlanetDef.dimension -> MOON

Hollow-core traversal (dig through to the other side):
- MoonCoreTraversalHandler: gravity state machine NORMAL(-Y) -> dig to core
  (Y<-38, falling) -> TRANSIT(+Y) + coordinate teleport (x,z)->(D-x,-z), D=100000
  -> dig up -> surface (Y>68) -> revert to NORMAL. Lands standing on the far side
- Moon low gravity (0.16g) + reduced fall damage

timescale + smoothness:
- /koniava timescale <0-100000> (SpaceTimeScalePacket S2C); accumulatedTime
  prevents position jump on rate change; worldPositionAt/combinedLightDir -> double

Sun/atmosphere fixes (from earlier this session):
- sun.fsh: bpodgursky procedural granulation + tempColor + quiescent loop
  prominences + rare eruptions (world-space angle, time-scaled)
- atmosphere.fsh: night-city-lights emissive additive on dark side;
  backlit overexposure fix (Mie 4->1.6 + clamp, forward-scatter edge ring)
- PlanetRenderer: night texture binding gated on locNightTex not locHasNight
- SpacePlanetManager.reload() no-op (AMD 26.x F3+T crash); per-frame texture
  upload cap

+++ src/main/java/com/github/nalamodikk/space/orbit/PlanetDef.java
@@ -25,13 +25,13 @@ public record PlanetDef(
     float    ringOuter,    // 外緣半徑 / 星球半徑
     float    ringTiltDeg   // 自轉軸傾斜度（土星 ≈ 26.7°）
 ) {
-    public Vector3f worldPositionAt(long gameTick, Vector3f starWorldPos) {
-        float progress = (gameTick / 24000.0f) / orbitalPeriodDays;
-        float angle    = (float) Math.toRadians(startAngleDeg + progress * 360.0f);
+    public Vector3f worldPositionAt(double gameTick, Vector3f starWorldPos) {
+        double progress = (gameTick / 24000.0) / orbitalPeriodDays;
+        double angle    = Math.toRadians(startAngleDeg + progress * 360.0);
         return new Vector3f(
-            starWorldPos.x + (float) Math.cos(angle) * orbitalRadius,
+            starWorldPos.x + (float) (Math.cos(angle) * orbitalRadius),
             starWorldPos.y,
-            starWorldPos.z + (float) Math.sin(angle) * orbitalRadius
+            starWorldPos.z + (float) (Math.sin(angle) * orbitalRadius)
         );
     }
